Encapsulate `setSingleRowMode` into conditional compilation to be able to compile against PostgreSQL 9.1
PostgreSQL 9.1 is the last version running on macOS using the libiconv version supplied by Apple, starting with 9.2 PostgreSQL requires a newer version of libiconv which causes problems due to the complex build of postgresql-libbpq from .hsc files and Apples tight security with shared libraries.

postgresql-libpq 0.9.2.0 added `setSingleRowMode` which made it impossible to compile against PostgreSQL 9.1.

Cabal 2.0 requires a change to Setup.hs, which was implemented in postgresql-libpq 0.9.3.1, leaving older versions unable to be compiled.

Since we need use Cabal 2.0 making postgresql-libpq 0.9.3.1 compatible with PostgreSQL 9.1 again was the most plausible way to go.